+++ test/benchmark-common-tasks/main.cpp
@@ -17,6 +17,7 @@ using namespace realm::util;
 using namespace realm::test_util;
 
 namespace {
+#define BASE_SIZE 3600
 
 /**
   This bechmark suite represents a number of common use cases,
@@ -34,8 +35,8 @@ namespace {
 
 const size_t min_repetitions = 10;
 const size_t max_repetitions = 1000;
-const double min_duration_s = 0.05;
-const double min_warmup_time_s = 0.01;
+const double min_duration_s = 0.1;
+const double min_warmup_time_s = 0.05;
 
 struct Benchmark
 {
@@ -91,8 +92,8 @@ struct BenchmarkWithStrings : BenchmarkWithStringsTable {
         BenchmarkWithStringsTable::before_all(group);
         WriteTransaction tr(group);
         TableRef t = tr.get_table("StringOnly");
-        t->add_empty_row(REALM_MAX_BPNODE_SIZE * 4);
-        for (size_t i = 0; i < REALM_MAX_BPNODE_SIZE * 4; ++i) {
+        t->add_empty_row(BASE_SIZE * 4);
+        for (size_t i = 0; i < BASE_SIZE * 4; ++i) {
             std::stringstream ss;
             ss << rand();
             t->set_string(0, i, ss.str());
@@ -110,9 +111,9 @@ struct BenchmarkWithLongStrings : BenchmarkWithStrings {
         t->insert_empty_row(0);
         // This should be enough to upgrade the entire array:
         t->set_string(0, 0, "A really long string, longer than 63 bytes at least, I guess......");
-        t->set_string(0, REALM_MAX_BPNODE_SIZE, "A really long string, longer than 63 bytes at least, I guess......");
-        t->set_string(0, REALM_MAX_BPNODE_SIZE * 2, "A really long string, longer than 63 bytes at least, I guess......");
-        t->set_string(0, REALM_MAX_BPNODE_SIZE * 3, "A really long string, longer than 63 bytes at least, I guess......");
+        t->set_string(0, BASE_SIZE, "A really long string, longer than 63 bytes at least, I guess......");
+        t->set_string(0, BASE_SIZE * 2, "A really long string, longer than 63 bytes at least, I guess......");
+        t->set_string(0, BASE_SIZE * 3, "A really long string, longer than 63 bytes at least, I guess......");
         tr.commit();
     }
 };
@@ -140,9 +141,9 @@ struct BenchmarkWithInts : BenchmarkWithIntsTable {
         BenchmarkWithIntsTable::before_all(group);
         WriteTransaction tr(group);
         TableRef t = tr.get_table("IntOnly");
-        t->add_empty_row(REALM_MAX_BPNODE_SIZE * 4);
+        t->add_empty_row(BASE_SIZE * 4);
         Random r;
-        for (size_t i = 0; i < REALM_MAX_BPNODE_SIZE * 4; ++i) {
+        for (size_t i = 0; i < BASE_SIZE * 4; ++i) {
             t->set_int(0, i, r.draw_int<int64_t>());
         }
         tr.commit();
@@ -181,6 +182,17 @@ struct BenchmarkSort : BenchmarkWithStrings {
         ConstTableRef table = tr.get_table("StringOnly");
         ConstTableView view = table->get_sorted_view(0);
     }
+};
+
+struct BenchmarkEmptyCommit : Benchmark {
+    const char* name() const { return "EmptyCommit"; }
+
+    void operator()(SharedGroup& group)
+    {
+        WriteTransaction tr(group);
+        tr.commit();
+    }
+
 };
 
 struct BenchmarkSortInt : BenchmarkWithInts {
@@ -357,10 +369,10 @@ struct BenchmarkGetLinkList : Benchmark {
 const char* durability_level_to_cstr(SharedGroup::DurabilityLevel level)
 {
     switch (level) {
-        case SharedGroup::durability_Full: return "Full";
+        case SharedGroup::durability_Full:    return "Full   ";
         case SharedGroup::durability_MemOnly: return "MemOnly";
 #ifndef _WIN32
-        case SharedGroup::durability_Async: return "Async";
+        case SharedGroup::durability_Async:   return "Async  ";
 #endif
     }
     return nullptr;
@@ -387,8 +399,14 @@ void run_benchmark(BenchmarkResults& results)
 {
     typedef std::pair<SharedGroup::DurabilityLevel, const char*> config_pair;
     std::vector<config_pair> configs;
-    configs.push_back(config_pair(SharedGroup::durability_Full, nullptr));
+
     configs.push_back(config_pair(SharedGroup::durability_MemOnly, nullptr));
+#if REALM_ENABLE_ENCRYPTION
+    configs.push_back(config_pair(SharedGroup::durability_MemOnly, crypt_key(true)));
+#endif
+
+    configs.push_back(config_pair(SharedGroup::durability_Full, nullptr));
+
 #if REALM_ENABLE_ENCRYPTION
     configs.push_back(config_pair(SharedGroup::durability_Full, crypt_key(true)));
 #endif
@@ -456,6 +474,7 @@ void run_benchmark(BenchmarkResults& results)
 
         results.finish(ident, lead_text_ss.str());
     }
+    std::cout << std::endl;
 }
 
 } // anonymous namespace
@@ -467,6 +486,7 @@ int benchmark_common_tasks_main()
     std::string results_file_stem = test_util::get_test_path_prefix() + "results";
     BenchmarkResults results(40, results_file_stem.c_str());
 
+    run_benchmark<BenchmarkEmptyCommit>(results);
     run_benchmark<AddTable>(results);
     run_benchmark<BenchmarkQuery>(results);
     run_benchmark<BenchmarkQueryNot>(results);
@@ -475,7 +495,9 @@ int benchmark_common_tasks_main()
     run_benchmark<BenchmarkSortInt>(results);
     run_benchmark<BenchmarkInsert>(results);
     run_benchmark<BenchmarkGetString>(results);
+
     run_benchmark<BenchmarkSetString>(results);
+
     run_benchmark<BenchmarkCreateIndex>(results);
     run_benchmark<BenchmarkGetLongString>(results);
     run_benchmark<BenchmarkSetLongString>(results);